Extract UserDashboardActivator from the monolithic Activator (#563 Sprint 7, A5)
Second per-module commit of the activation split: the user-dashboard module's
activation work moves to a new
FreeFormCertificate\UserDashboard\UserDashboardActivator::create_tables() —
the base ffc_user role + administrator capability grant, the ffc_user_profiles
and ffc_custom_fields tables, and the front-end "My Dashboard" page. The core
Activator now delegates via the same class_exists guard used for the other
modules. Behavior-preserving — ActivatorTest / UserManagerTest green;
activation order unchanged.
